+++ api/src/controllers/client/index.ts
@@ -1,5 +1,8 @@
 // Copyright 2023 Touca, Inc. Subject to Apache-2.0 License.
 
+export { clientAuthTokenCreate } from './authTokenCreate.js'
+export { clientAuthTokenStatus } from './authTokenStatus.js'
+export { clientAuthTokenVerify } from './authTokenVerify.js'
 export { clientBatchNext } from './batchNext.js'
 export { clientElementList } from './elementList.js'
 export { clientOptions } from './options.js'