+++ app/src/app/app.component.ts
@@ -1,4 +1,4 @@
-// Copyright 2021 Touca, Inc. Subject to Apache-2.0 License.
+// Copyright 2023 Touca, Inc. Subject to Apache-2.0 License.
 
 import { Component } from '@angular/core';
 import { Title } from '@angular/platform-browser';