+++ app/src/app/app.module.ts
@@ -1,4 +1,4 @@
-// Copyright 2021 Touca, Inc. Subject to Apache-2.0 License.
+// Copyright 2023 Touca, Inc. Subject to Apache-2.0 License.
 
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';