+++ app/src/app/core/services/guest.guard.ts
@@ -1,4 +1,4 @@
-// Copyright 2022 Touca, Inc. Subject to Apache-2.0 License.
+// Copyright 2023 Touca, Inc. Subject to Apache-2.0 License.
 
 import { Injectable } from '@angular/core';
 import {
@@ -6,11 +6,11 @@ import {
   CanActivate,
   CanActivateChild,
   Router,
-  RouterStateSnapshot,
-  UrlTree
+  RouterStateSnapshot
 } from '@angular/router';
 import { map } from 'rxjs';
 
+import { ELocalStorageKey } from '../models/frontendtypes';
 import { ApiService } from './api.service';
 import { AuthService } from './auth.service';
 
@@ -31,6 +31,12 @@ export class GuestGuard implements CanActivate, CanActivateChild {
           return this.router.parseUrl('/account/install');
         }
         if (this.authService.isLoggedIn()) {
+          if (route.queryParamMap.has('t')) {
+            localStorage.setItem(
+              ELocalStorageKey.CLIToken,
+              route.queryParamMap.get('t')
+            );
+          }
           this.authService.redirectUrl = state.url;
           return this.router.parseUrl('/~');
         }